(PC-HT) move get_adage_offerer into AdageClient
