+++ api/src/pcapi/core/educational/adage_backends/base.py
@@ -1,4 +1,5 @@
 from pcapi import settings
+from pcapi.connectors.serialization.api_adage_serializers import AdageVenue
 from pcapi.core.educational.models import AdageApiResult
 from pcapi.routes.adage.v1.serialization.prebooking import EducationalBookingResponse
 
@@ -8,4 +9,7 @@ class AdageClient:
         self.base_url = settings.ADAGE_API_URL
 
     def notify_prebooking(self, data: EducationalBookingResponse) -> AdageApiResult:
+        raise NotImplementedError()
+
+    def get_adage_offerer(self, siren: str) -> list[AdageVenue]:
         raise NotImplementedError()
